Fix headers and add "testing" bypass header
Fix the header name passed down and add in the super-secret header
for initial testing purposes (so services know to accept from the
LB the passed in headers). This secret header/code combination is
mirrored in a minor change to keystonemiddleware.

Long term this should only ever occur with trusted certs between
the middleware and the services.